+++ dubbo-samples-callback/src/main/java/org/apache/dubbo/samples/callback/CallbackProvider.java
@@ -17,13 +17,20 @@
  *
  */
 
-package com.alibaba.dubbo.samples.callback.api;
+package org.apache.dubbo.samples.callback;
+
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 /**
- * CallbackListener
+ * MergeProvider
  */
-public interface CallbackListener {
+public class CallbackProvider {
 
-    void changed(String msg);
+    public static void main(String[] args) throws Exception {
+        new EmbeddedZooKeeper(2181, false).start();
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/callback-provider.xml"});
+        context.start();
+        System.in.read();
+    }
 
 }
