+++ dubbo-samples-callback/src/main/java/org/apache/dubbo/samples/callback/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.callback;
+package org.apache.dubbo.samples.callback;
 
 import java.io.File;
 import java.lang.reflect.Method;
