+++ dubbo-samples-callback/src/main/java/org/apache/dubbo/samples/callback/api/CallbackListener.java
@@ -17,20 +17,13 @@
  *
  */
 
-package com.alibaba.dubbo.samples.callback;
-
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+package org.apache.dubbo.samples.callback.api;
 
 /**
- * MergeProvider
+ * CallbackListener
  */
-public class CallbackProvider {
+public interface CallbackListener {
 
-    public static void main(String[] args) throws Exception {
-        new EmbeddedZooKeeper(2181, false).start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/callback-provider.xml"});
-        context.start();
-        System.in.read();
-    }
+    void changed(String msg);
 
 }
