+++ dubbo-samples-callback/src/main/resources/spring/callback-consumer.xml
@@ -29,6 +29,6 @@
     <!--<dubbo:registry address="multicast://224.5.6.7:1234"/>-->
     <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
 
-    <dubbo:reference id="callbackService" interface="com.alibaba.dubbo.samples.callback.api.CallbackService"/>
+    <dubbo:reference id="callbackService" interface="org.apache.dubbo.samples.callback.api.CallbackService"/>
 
 </beans>
